refactor(Ticker.jsx): Reorganizing files in the directory
