+++ app/frontend/features/tickers/Ticker.jsx
